+++ shellish/command.py
@@ -196,8 +196,8 @@ class Command(object):
     def tabulate(self, *args, **kwargs):
         return layout.tabulate(*args, **kwargs)
 
-    def vtprint(self, *args, **kwargs):
-        return layout.vtprint(*args, **kwargs)
+    def vtmlprint(self, *args, **kwargs):
+        return layout.vtmlprint(*args, **kwargs)
 
     def tree(self, *args, **kwargs):
         return layout.dicttree(*args, **kwargs)